[RefData][Geonames] Now spot the entries with duplicated IATA codes when extracting POR details.

+++ refdata/geonames/data/por/admin/extract_por_with_iata_icao.sh
@@ -5,27 +5,31 @@
 # - the port of the database
 #
 
-if [ "$1" = "-h" -o "$1" = "--help" ];
+##
+# Temporary path
+TMP_DIR="/tmp/por"
+
+##
+# Path of the executable: set it to empty when this is the current directory.
+EXEC_PATH=`dirname $0`
+CURRENT_DIR=`pwd`
+if [ ${CURRENT_DIR} -ef ${EXEC_PATH} ]
 then
-	echo "Usage: $0 [<Database Server Hostname> [<Database Server Port>]]"
-	echo ""
-	exit -1
+	EXEC_PATH="."
+	TMP_DIR="."
+fi
+EXEC_PATH="${EXEC_PATH}/"
+TMP_DIR="${TMP_DIR}/"
+
+if [ ! -d ${TMP_DIR} -o ! -w ${TMP_DIR} ]
+then
+	\mkdir -p ${TMP_DIR}
 fi
 
 ##
-# Database Server Hostname
+# Database parameters
 DB_HOST="localhost"
-if [ "$1" != "" ];
-then
-	DB_HOST="$1"
-fi
-
-# Database Server Port
 DB_PORT="3306"
-if [ "$2" != "" ];
-then
-	DB_PORT="$2"
-fi
 
 # Database User
 DB_USER="geo"
@@ -38,13 +42,61 @@ DB_NAME="geo_geonames"
 
 # Snapshot date
 SNAPSHOT_DATE=`date "+%Y%m%d"`
+SNAPSHOT_DATE_HUMAN=`date`
 
+##
 # Extract airport/city information from the Geonames tables (in particular,
-# geoname and alternate_name)
-SQL_FILE="extract_por_with_iata_icao.sql"
-DUMP_FILE="por_all_iata_${SNAPSHOT_DATE}.csv"
-DUMP_FILE_ICAO_ONLY="por_all_icao_only_${SNAPSHOT_DATE}.csv"
-DUMP_FILE_NO_CODE="por_all_nocode_${SNAPSHOT_DATE}.csv"
+# 'geoname' and 'alternate_name')
+SQL_FILE_FILENAME=extract_por_with_iata_icao.sql
+SQL_FILE=${EXEC_PATH}${SQL_FILE_FILENAME}
+# Generated files
+DUMP_FILE_FILENAME=por_all_iata_${SNAPSHOT_DATE}.csv
+DUMP_FILE_ICAO_ONLY_FILENAME=por_all_icao_only_${SNAPSHOT_DATE}.csv
+DUMP_FILE_NO_CODE_FILENAME=por_all_nocode_${SNAPSHOT_DATE}.csv
+DUMP_FILE_NO_ICAO_FILENAME=por_all_noicao_${SNAPSHOT_DATE}.csv
+DUMP_FILE_DUP_FILENAME=por_all_dup_iata_${SNAPSHOT_DATE}.csv
+#
+DUMP_FILE=${TMP_DIR}${DUMP_FILE_FILENAME}
+DUMP_FILE_ICAO_ONLY=${TMP_DIR}${DUMP_FILE_ICAO_ONLY_FILENAME}
+DUMP_FILE_NO_CODE=${TMP_DIR}${DUMP_FILE_NO_CODE_FILENAME}
+DUMP_FILE_NO_ICAO=${TMP_DIR}${DUMP_FILE_NO_ICAO_FILENAME}
+DUMP_FILE_DUP=${TMP_DIR}${DUMP_FILE_DUP_FILENAME}
+
+#
+if [ "$1" = "-h" -o "$1" = "--help" ];
+then
+	echo
+	echo "Usage: $0 [<Database Server Hostname> [<Database Server Port>]]"
+	echo "  - Default database server hostname: '${DB_HOST}'"
+	echo "  - Default database server port: '${DB_PORT}'"
+	echo "  - Database username: '${DB_USER}'"
+	echo "  - Database name: '${DB_NAME}'"
+	echo "  - Snapshot date: '${SNAPSHOT_DATE}' (${SNAPSHOT_DATE_HUMAN})"
+	echo "  - Generated (CSV-formatted) data files:"
+	echo "      + '${DUMP_FILE}'"
+	echo "      + '${DUMP_FILE_ICAO_ONLY}'"
+	echo "      + '${DUMP_FILE_NO_CODE}'"
+	echo "      + '${DUMP_FILE_NO_ICAO}'"
+	echo "      + '${DUMP_FILE_DUP}'"
+	echo
+	exit -1
+fi
+
+##
+# Database Server Hostname
+if [ "$1" != "" ];
+then
+	DB_HOST="$1"
+fi
+
+# Database Server Port
+if [ "$2" != "" ];
+then
+	DB_PORT="$2"
+fi
+
+##
+#
 echo
 echo "Exporting points of reference (POR, i.e., airports, cities) data from the tables of Geonames into '${DUMP_FILE}'."
 echo "That operation may take several minutes..."
@@ -52,7 +104,9 @@ mysql -u ${DB_USER} --password=${DB_PASSWD} -P ${DB_PORT} -h ${DB_HOST} ${DB_NAM
 echo "... Done"
 echo
 
-# Remove the first line (header)
+##
+# Remove the first line (header). Note: that step should now be performed by
+# the caller.
 #sed -i -e "s/^code\(.\+\)//g" ${DUMP_FILE}
 #sed -i -e "/^$/d" ${DUMP_FILE}
 
@@ -73,8 +127,9 @@ grep "^NULL\^\(.\+\)" ${DUMP_FILE} > ${DUMP_FILE_ICAO_ONLY}
 sed -i -e "s/^NULL\^\(.\+\)//g" ${DUMP_FILE}
 sed -i -e "/^$/d" ${DUMP_FILE}
 
-# 3.0. We are now left with only the points of interest containing a non-NULL
-#      IATA code.
+##
+# We are now left with only the points of interest containing a non-NULL IATA
+# code.
 
 # 3.1. Extract the header into a temporary file
 DUMP_FILE_HEADER=${DUMP_FILE}.tmp.hdr
@@ -84,18 +139,66 @@ grep "^alternateName\(.\+\)" ${DUMP_FILE} > ${DUMP_FILE_HEADER}
 sed -i -e "s/^alternateName\(.\+\)//g" ${DUMP_FILE}
 sed -i -e "/^$/d" ${DUMP_FILE}
 
-# 3.3. Remove duplicated entries for IATA codes set for cities (e.g, AMS, LAX).
-#      First, the NULL ICAO code is replaced by aaaa, so that it appears on the
-#      second line, when the IATA code is duplicated. Hence, with the uniq
-#      command, that second line will be deleted.
-DUMP_UNIQ_FILE=${DUMP_FILE}.tmp.uniq
-sed -i -e "s/\(.\+\)\^NULL\^\(.\+\)/\1\^aaaa\^\2/g" ${DUMP_FILE}
-sort -t '^' -k1,1 -k2,2r ${DUMP_FILE} | uniq -w 3 > ${DUMP_UNIQ_FILE}
+# 4.1. Extract the entries having no ICAO code.
+grep "^\([A-Z][A-Z][A-Z]\)\^NULL\^\(.\+\)" ${DUMP_FILE} > ${DUMP_FILE_NO_ICAO}
 
-# 3.4. Re-add the header
+# 4.2. Remove the entries having no ICAO code.
+#      Note that there is no ICAO code for a city. Hence, city entries are also
+#      filtered out.
+#      Note also that, when the same IATA code is used for a city and one of its
+#      airports (e.g, AMS, LAX, SFO), two entries with the same IATA code
+#      should appear. But, as the city entries are removed in this step, that
+#      case of IATA duplicity is avoided/removed.
+sed -i -e "s/^\([A-Z][A-Z][A-Z]\)\^NULL\^\(.\+\)//g" ${DUMP_FILE}
+
+# 4.3. Spot the (potential) remaining entries having duplicated IATA codes.
+#      Here, only the airport entries having duplicated IATA codes are spotted.
+#      That case may typically appear when someone, in Geonames, has mistakenly
+#      set the IATA code (say ACQ; that airport is Waseca Municpal Airport and
+#      its ICAO code is KACQ) in place of the FAA code (indeed, ACQ is the FAA
+#      code, not the IATA code).
+#
+#      With the uniq command, all the entries having a duplicated IATA code are
+#      deleted. Then, the original file (with potential duplicated entries) is
+#      compared with the de-duplicated file: the differences are the duplicated
+#      entries.
+#
+# 4.3.1. Create the file with no duplicated IATA code.
+DUMP_UNIQ_FILE=${DUMP_FILE}.tmp.uniq
+DUMP_FILE_TMP=${DUMP_FILE}.tmp
+sort -t '^' -k1,1 -k2,2 ${DUMP_FILE} > ${DUMP_FILE_TMP}
+\mv -f ${DUMP_FILE_TMP} ${DUMP_FILE}
+uniq -w 3 ${DUMP_FILE} > ${DUMP_UNIQ_FILE}
+
+# 4.3.2. Create the file with only the duplicated IATA code entries, if any.
+DUMP_FILE_DUP_CHECK=${DUMP_FILE_DUP}.tmp.check
+comm -23 ${DUMP_FILE} ${DUMP_UNIQ_FILE} > ${DUMP_FILE_DUP_CHECK}
+sed -i -e "/^$/d" ${DUMP_FILE_DUP_CHECK}
+
+if [ -s ${DUMP_FILE_DUP_CHECK} ]
+then
+	POR_DUP_IATA_NB=`wc -l ${DUMP_FILE_DUP_CHECK} | cut -d' ' -f1`
+	echo
+	echo "!!!!!! WARNING !!!!!!!!"
+	echo "Geonames has got ${POR_DUP_IATA_NB} duplicated IATA codes (in addition to those of cities of course). To see them, just do:"
+	echo "less ${DUMP_FILE_DUP_CHECK}"
+	echo "Note: they result of the comparison between '${DUMP_FILE}' (all POR) and"
+	echo "'${DUMP_UNIQ_FILE}' (duplicated POR have been removed)."
+	echo "!!!!!! WARNING !!!!!!!!"
+	echo
+else
+	\rm -f ${DUMP_FILE_DUP_CHECK}
+fi
+
+# 4.4. Re-add the header
 cat ${DUMP_FILE_HEADER} ${DUMP_UNIQ_FILE} > ${DUMP_FILE}
-sed -i -e "s/\(.\+\)\^aaaa\^\(.\+\)/\1\^NULL\^\2/g" ${DUMP_FILE}
-\rm -f ${DUMP_FILE_HEADER} ${DUMP_UNIQ_FILE}
+
+##
+# Clean
+if [ "${TMP_DIR}" != "/tmp/por/" ]
+then
+	\rm -f ${DUMP_FILE_HEADER} ${DUMP_UNIQ_FILE}
+fi
 
 
 # Replace the NULL fields by empty fields
